refactor(tls): extract verifier tests to sibling file
